Add get_followers
